requirements.txt
click==7.1.2
python-dotenv>=0.5.1
scikit-learn==0.24.1
dataclasses==0.6
pyyaml==6.0
marshmallow-dataclass==8.3.0
pandas==1.1.5
boto3==1.17.49
mlflow==1.15.0
pytest==7.2.0
nbformat==5.1.3
hydra-core==1.2.0
omegaconf==2.2.3
opencv-python==4.6.0.66
protobuf==3.20.0